# api/openapi/control-v1alpha1.yaml
openapi: 3.1.0
info:
  title: Agent OS Control API
  version: v1alpha1
  description: Public task submission and observation contract. Task status is kernel-owned.
servers:
  - url: https://control.agentos.dev
paths:
...
  /v1/tools:
    get:
      operationId: listTools
      summary: Discover the tools the authenticated tenant may invoke
      description: >-
        Returns the versioned tool descriptors registered in the tenant's
        registry. Invocation is gated by policy outside the LLM; this endpoint
        only lists declared capabilities.
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: The tenant's registered tool descriptors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolList'
        '401': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: Structured request failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    ToolDescriptor:
      type: object
      additionalProperties: false
      required: [name, version, sideEffectRisk, actions, resourcePatterns, paramsSchema, specDigest, createdAt]
      properties:
        name: { type: string }
        version: { type: string }
        sideEffectRisk:
          type: string
          enum: [none, low, high]
        actions:
          type: array
          items: { type: string }
        resourcePatterns:
          type: array
          items: { type: string }
        paramsSchema:
          type: object
          description: JSON Schema the invocation arguments must satisfy.
        specDigest:
          type: string
          description: SHA-256 of the immutable descriptor spec.
        createdAt: { type: string, format: date-time }
    ToolList:
      type: object
      additionalProperties: false
      required: [apiVersion, kind, tools, traceId]
      properties:
        apiVersion: { type: string, const: 'agentos.dev/v1alpha1' }
        kind: { type: string, const: ToolList }
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ToolDescriptor'
        traceId: { type: string, format: uuid }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, reasonCode, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        reasonCode: { type: string }
        traceId: { type: string, format: uuid }
security:
  - bearerAuth: []